Refactor: 폴더링 변경, 컴포넌트 분리
