+++ src/feature/post/ui/RenderPostTable.tsx
@@ -2,87 +2,27 @@ import { Edit2, MessageSquare, ThumbsDown, ThumbsUp, Trash2 } from "lucide-react
 import { Table, TableBody, TableCell, TableHead, TableHeader, TableRow } from "../../../shared/ui/table/Table"
 import { highlightText } from "../../../shared/utils/highlightText"
 import { Button } from "../../../shared/ui/button/Button"
-import {
-  postsAtom,
-  searchQueryAtom,
-  selectedPostAtom,
-  selectedTagAtom,
-  showPostDetailDialogAtom,
-  showUserModalAtom,
-} from "../model/postAtoms"
+import { postsAtom, searchQueryAtom, selectedTagAtom } from "../model/postAtoms"
 
 import { useAtom } from "jotai"
 import { useUpdateURL } from "../../../shared/model/urlUtils"
 
-import { userAtom } from "../../../entities/model/atom"
-import { commentFetch } from "../../comment/model/commentFetch"
-import { commentsAtom } from "../../comment/model/commentAtom"
-import { usePost } from "../model/usePost"
 import { usePostHandler } from "../model/postHandler"
-import { userFetchData } from "../../../entities/model/userFetch"
-import { Post } from "../model/postType"
+import { usePostDialog } from "../model/usePostDialog"
 
 // 게시물 테이블
 export const RenderPostTable = () => {
-  const [, setUser] = useAtom(userAtom)
-  const [posts, setPosts] = useAtom(postsAtom)
+  const [posts] = useAtom(postsAtom)
   const [searchQuery] = useAtom(searchQueryAtom)
   const [selectedTag, setSelectedTag] = useAtom(selectedTagAtom)
-  const [, setShowUserModal] = useAtom(showUserModalAtom)
-  const [, setShowPostDetailDialog] = useAtom(showPostDetailDialogAtom)
-  const [, setComments] = useAtom(commentsAtom)
-  const [, setSelectedPost] = useAtom(selectedPostAtom)
 
+  const { handleDeletePost } = usePostHandler()
+  const { handleOpenPostUpdate, handleOpenPostDetail, handleOpenUserDetail } = usePostDialog()
   const updateURL = useUpdateURL()
-
-  const { deletePost } = usePost()
 
   const handleTagClick = (tag: string) => {
     setSelectedTag(tag)
     updateURL()
-  }
-
-  const handleOpenUserModal = async (userId: number) => {
-    try {
-      const userData = await userFetchData(userId)
-
-      setUser(userData)
-      setShowUserModal(true)
-    } catch (error) {
-      console.error("사용자 데이터를 가져오는 중 오류 발생:", error)
-    }
-  }
-
-  const handleOpenPostDetail = async (post: Post) => {
-    setSelectedPost(post)
-    try {
-      const getComment = await commentFetch(post.id)
-      console.log("getComment", getComment)
-      if (!getComment) {
-        return
-      }
-
-      setComments((prev) => ({ ...prev, [post.id]: getComment }))
-      setShowPostDetailDialog(true)
-    } catch (error) {
-      console.error("댓글을 가져오는 중 오류 발생:", error)
-    }
-  }
-  const { handleOpenPostUpdate } = usePostHandler()
-
-  const handleDeletePost = async (postId: number) => {
-    try {
-      const deleteResponse = await deletePost(postId)
-      if (deleteResponse) {
-        setPosts((prevPosts) => prevPosts.filter((post) => post.id !== postId))
-        alert("삭제 완료")
-      } else {
-        alert("삭제 오류")
-      }
-    } catch (error) {
-      console.error("게시글 삭제 중 오류 발생:", error)
-      alert("삭제 중 오류 발생")
-    }
   }
 
   return (
@@ -123,7 +63,7 @@ export const RenderPostTable = () => {
             <TableCell>
               <div
                 className="flex items-center space-x-2 cursor-pointer"
-                onClick={() => post.author?.id && handleOpenUserModal(post.author.id)}
+                onClick={() => post.author?.id && handleOpenUserDetail(post.author.id)}
               >
                 <img src={post.author?.image} alt={post.author?.username} className="w-8 h-8 rounded-full" />
                 <span>{post.author?.username}</span>
